reorganize k8 objects and controller into separate modules
