+++ src/sc/src/k8/objects/mod.rs
@@ -1,0 +1,5 @@
+pub mod spg_group;
+pub mod spg_service;
+pub mod spu_k8_config;
+pub mod statefulset;
+pub mod spu_service;
