+++ src/sc/src/k8/objects/spu_k8_config.rs
