+++ src/sc/src/k8/objects/statefulset.rs
