--- src/sc/src/k8/service/mod.rs
@@ -1,5 +1,0 @@
-mod controller;
-mod meta;
-
-pub use self::meta::*;
-pub use self::controller::SpuServiceController;
